PA_Backend/pyproject.toml
[tool.poetry]
name = "pa-backend"
version = "0.1.0"
description = "AI Agent Orchestration Server"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
accelerate = "1.7.0"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.12.1"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.9.0"
attrs = "25.3.0"
cachetools = "5.5.2"
certifi = "2025.4.26"
charset-normalizer = "3.4.2"
click = "8.2.1"
colorama = "0.4.6"
dataclasses-json = "0.6.7"
dotenv = "0.9.9"
faiss-cpu = "1.11.0"
fastapi = "0.115.12"
filelock = "3.18.0"
frozenlist = "1.6.0"
fsspec = "2025.5.1"
future = "1.0.0"
gkeepapi = "0.16.0"
google-api-core = "2.24.2"
google-api-python-client = "2.170.0"
google-auth = "2.40.2"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.2"
googleapis-common-protos = "1.70.0"
gpsoauth = "1.1.1"
greenlet = "3.2.2"
h11 = "0.16.0"
httpcore = "1.0.9"
httplib2 = "0.22.0"
httptools = "0.6.4"
httpx = "0.28.1"
httpx-sse = "0.4.0"
huggingface-hub = "0.32.1"
HyperPyYAML = "1.2.2"
idna = "3.10"
Jinja2 = "3.1.6"
joblib = "1.5.1"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
langchain = "0.3.25"
langchain-community = "0.3.24"
langchain-core = "0.3.61"
langchain-huggingface = "0.2.0"
langchain-text-splitters = "0.3.8"
langgraph = "0.4.7"
langgraph-checkpoint = "2.0.26"
langgraph-prebuilt = "0.2.1"
langgraph-sdk = "0.1.70"
langsmith = "0.3.42"
MarkupSafe = "3.0.2"
marshmallow = "3.26.1"
mpmath = "1.3.0"
multidict = "6.4.4"
mypy-extensions = "*"
networkx = "3.4.2"
numpy = "2.2.6"
oauthlib = "3.2.2"
orjson = "3.10.18"
ormsgpack = "1.10.0"
packaging = "24.2"
pillow = "11.2.1"
propcache = "0.3.1"
proto-plus = "1.26.1"
protobuf = "6.31.0"
psutil = "7.0.0"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.2"
pycryptodomex = "3.23.0"
pydantic = "2.11.5"
pydantic-settings = "2.9.1"
pydantic_core = "2.33.2"
pyparsing = "3.2.3"
python-dotenv = "1.1.0"
PyYAML = "6.0.2"
regex = "2024.11.6"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
requests-toolbelt = "1.0.0"
rsa = "4.9.1"
ruamel-yaml = "0.18.11"
ruamel-yaml-clib = "0.2.12"
safetensors = "0.5.3"
scikit-learn = "1.6.1"
scipy = "1.15.3"
sentence-transformers = "4.1.0"
sentencepiece = "0.2.0"
setuptools = "80.8.0"
sniffio = "1.3.1"
speechbrain = "1.0.3"
SQLAlchemy = "2.0.41"
starlette = "0.46.2"
sympy = "1.14.0"
tenacity = "9.1.2"
threadpoolctl = "3.6.0"
tokenizers = "0.21.1"
tqdm = "4.67.1"
typing-inspect = "0.9.0"
typing-inspection = "0.4.1"
typing_extensions = "4.13.2"
uritemplate = "4.1.1"
urllib3 = "1.26.20"
uvicorn = "0.34.2"
watchfiles = "1.0.5"
websockets = "15.0.1"
xxhash = "3.5.0"
yarl = "1.20.0"
zstandard = "0.23.0"
psycopg2-binary = "^2.9.6"
sqlmodel = "*"
asyncpg = "*"
alembic = "^1.11.1"
pandas = "^2.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^7.4.0"
requests = "^2.31.0"

[tool.poetry.scripts]
lint = "flake8 ."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"